fix quite a few warnings that occur on clang when using -Weverything
most warnings were in DUNE and ALUGrid, but these have been
ignored. Also fixing some of these warnings (in particular the
"parameter foo is unused" ones) would make the code harder to read and
understand, so they have been ignored, too...

+++ ewoms/io/vtkmultiwriter.hh
@@ -156,7 +156,6 @@ public:
     /*!
      * \brief Add a finished vertex centered vector field to the
      *        output.
-     * \brief Add a vertex-centered quantity to the output.
      *
      * If the buffer is managed by the VtkMultiWriter, it must have
      * been created using createField() and may not be used by
@@ -280,7 +279,7 @@ public:
             if (fileLen > 0) {
                 std::ifstream multiFileIn(multiFileName_.c_str());
                 char *tmp = new char[fileLen];
-                multiFileIn.read(tmp, fileLen);
+                multiFileIn.read(tmp, static_cast<long>(fileLen));
                 res.serializeStream().write(tmp, fileLen);
                 delete[] tmp;
             }
